services/service-c-climate/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pymongo>=4.6.0
pandas>=2.1.0
python-dotenv>=1.0.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.27.0